final (beta)

+++ main.py
@@ -1,14 +1,14 @@
 import asyncio
 import sqlite3 as sl
 from aiogram import Bot, Dispatcher, types, F
-from aiogram.filters import Command, CommandObject # импортируем токен
+from aiogram.filters import Command, CommandObject  # импортируем токен
 import logging
 from aiogram.types import ReplyKeyboardRemove, ReplyKeyboardMarkup, KeyboardButton
 
 # Запускаем логгирование
-# logging.basicConfig(
-#     format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.DEBUG
-# )
+logging.basicConfig(
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.DEBUG
+)
 
 # создаем диспетчер
 bot = Bot(token='7814578784:AAFEIg8-DKS1z6Lte7VGGa_lykNc1xy6GaM')
@@ -21,7 +21,7 @@ one = [[KeyboardButton(text='/+ баланс')],
 klava_one = ReplyKeyboardMarkup(keyboard=one, resize_keyboard=True, one_time_keyboard=True)
 
 two = [[KeyboardButton(text='/только_глобальный_рынок')],
-        [KeyboardButton(text='/с_китайский_рынком')],
+       [KeyboardButton(text='/с_китайский_рынком')],
        [KeyboardButton(text='/назад')]]
 klava_two = ReplyKeyboardMarkup(keyboard=two, resize_keyboard=True, one_time_keyboard=True)
 
@@ -29,31 +29,26 @@ three = [[KeyboardButton(text='/только_новые')],
          [KeyboardButton(text='/можно и б\у')],
          [KeyboardButton(text='/назад')]]
 klava_three = ReplyKeyboardMarkup(keyboard=three, resize_keyboard=True, one_time_keyboard=True)
-four = [[KeyboardButton(text='/- apple')],[KeyboardButton(text='/- samsung')],[KeyboardButton(text='/- xiaomi')],
-         [KeyboardButton(text='/- poco')],[KeyboardButton(text='/- oneplus')],[KeyboardButton(text='/- realme')],
-         [KeyboardButton(text='/- oppo')],[KeyboardButton(text='/- honor')], [KeyboardButton(text='/- vivo')],
-         [KeyboardButton(text='/- iqoo')], [KeyboardButton(text='/- pixel')],
-         [KeyboardButton(text='/далее')], [KeyboardButton(text='/назад')]]
+
+four = [[KeyboardButton(text='/- apple')], [KeyboardButton(text='/- samsung')], [KeyboardButton(text='/- xiaomi')],
+        [KeyboardButton(text='/- poco')], [KeyboardButton(text='/- oneplus')], [KeyboardButton(text='/- realme')],
+        [KeyboardButton(text='/- oppo')], [KeyboardButton(text='/- honor')], [KeyboardButton(text='/- vivo')],
+        [KeyboardButton(text='/- iqoo')], [KeyboardButton(text='/- pixel')], [KeyboardButton(text='/- huawei')],
+        [KeyboardButton(text='/далее')], [KeyboardButton(text='/назад')]]
 klava_four = ReplyKeyboardMarkup(keyboard=four, resize_keyboard=True, one_time_keyboard=True)
+
 start = [[KeyboardButton(text='/start')]]
 klava_start = ReplyKeyboardMarkup(keyboard=start, resize_keyboard=True, one_time_keyboard=True)
+
 end = [[KeyboardButton(text='/оставить_отзыв')], [KeyboardButton(text='/назад')]]
 klava_end = ReplyKeyboardMarkup(keyboard=end, resize_keyboard=True, one_time_keyboard=True)
 
-# phons = ['apple - 13 б/у, баланс', 'samsung - s23 китай, new, баланс-камерофон',
-#          'samsung - s23Fe баланс-камерофон', 'samsung - s22 Ultra б/у, баланс-камерофон',
-#          'xiaomi - 13T Pro баланс', 'xiaomi - 14T баланс-камерофон', 'xiaomi - 12S Ultra б/у, баланс-камерофон',
-#          'xiaomi - 11Ultra камерофон', 'xiaomi - 13 китай, new, баланс-камерофон',
-#          'xiaomi - K70 Ultra китай, new, игрофон', 'poco - X7 Pro игрофон', 'oneplus - 11 баланс-камерофон',
-#          'oneplus - Ace 5 китай, new, игрофон', 'realme - 14 Pro+ камерофон', 'realme - Neo 7 китай, new, игрофон',
-#          'realme - GT6 китай, б/у, игрофон', 'oppo - Find X6 Pro китай, б/у, баланс-камерофон', 'honor - 200 камерофон',
-#          'honor - 200 Pro китай, new, камерофон', 'honor - GT китай, new, игрофон',
-#          'vivo - X90 Pro+ китай, б/у, баланс-камерофон', 'iqoo - 12 китай, б/у, баланс',
-#          'iqoo - Neo 9s Pro+ китай, new, баланс', 'iqoo - Neo 10 китай, new, игрофон', 'pixel - 7 pro б/у, камерофон']
 exit = []
 delt = ''
 fd = False
 recomendation = 0
+KEY = 'amol'
+
 
 async def main():
     await bot.delete_webhook()
@@ -61,8 +56,10 @@ async def main():
 
 
 @dp.message(Command('start'))  # декоратор для обработчика команды start
-
 async def process_start_command(message: types.Message):
+    global recomendation, delt
+    recomendation = 0
+    delt = ''
     """
     Создаем и регистрируем в диспетчере асинхронный обработчик сообщений.
     В параметре message содержится вся информация о сообщении - см. документацию.
@@ -75,11 +72,11 @@ async def process_start_command(message: types.Message):
                         reply_markup=klava_one)  # отправляет ответ на сообщение
 
 
-
 @dp.message(Command('help'))  # декоратор для обработчика команды help
 async def process_help_command(message: types.Message):
     await message.reply("Привет!\nЯ помогу выбрать вам телефон в популярном ценовом диапозоне от 25000 до "
                         "30000 рублей\n\nМаксимум что вам надо писать в чат это команду /start, а остальное - кнопки:)")
+
 
 @dp.message(Command('назад'))
 async def out(message: types.Message):
@@ -88,6 +85,7 @@ async def out(message: types.Message):
     delt = ''
     recomendation = 0
     await message.reply('Вернулись назад :)\n/start', reply_markup=klava_start)
+
 
 @dp.message(Command('+'))
 async def balance(message: types.Message):
@@ -106,7 +104,9 @@ async def balance(message: types.Message):
                         'глобальную версию (система от глобальной версии со всеми языками и легкой настройкой, но '
                         'бэндов не прибавится). За все эти муки смартфоны обходятся по низкой цене :). Доставка так же '
                         'из Китая и если придет брак, то отправить его обратно будет не просто. Так же некоторые '
-                        'смартфоны в китайской версии имеют лучшую оболочку, чем в глобальной версии ;)', reply_markup=klava_two)
+                        'смартфоны в китайской версии имеют лучшую оболочку, чем в глобальной версии ;)',
+                        reply_markup=klava_two)
+
 
 @dp.message(Command('с_китайский_рынком'))
 async def china(message: types.Message):
@@ -121,6 +121,7 @@ async def china(message: types.Message):
                         'состоянии, или быть разбитым. Вы покупаете кота в мешке, но цена тут гораздо ниже, чем у новых,'
                         'а также есть прошлогодние флагманы по хорошей цене :)', reply_markup=klava_three)
 
+
 @dp.message(Command('только_глобальный_рынок'))
 async def globals(message: types.Message):
     global exit
@@ -134,6 +135,7 @@ async def globals(message: types.Message):
                         'состоянии, или быть разбитым. Вы покупаете кота в мешке, но цена тут гораздо ниже, чем '
                         'у новых, а также есть прошлогодние флагманы по хорошей цене :)', reply_markup=klava_three)
 
+
 @dp.message(Command('только_новые'))
 async def new(message: types.Message):
     global exit
@@ -144,6 +146,7 @@ async def new(message: types.Message):
     await message.reply('Теперь исключите все марки смартфонов которые вы не хотите видеть у себя в рекомендации,'
                         'а потом нажмите /далее и ваш список готов :). Если вы хотите увидеть всех - просто нажмите '
                         '/далее', reply_markup=klava_four)
+
 
 @dp.message(Command('можно'))
 async def use(message: types.Message):
@@ -156,6 +159,7 @@ async def use(message: types.Message):
                         'а потом нажмите /далее и ваш список готов :). Если вы хотите увидеть всех - просто нажмите '
                         '/далее', reply_markup=klava_four)
 
+
 @dp.message(Command('-'))
 async def delete(message: types.Message):
     global delt
@@ -167,9 +171,11 @@ async def delete(message: types.Message):
     else:
         await message.reply('Вы не написали марку', reply_markup=klava_four)
 
+
 @dp.message(Command('далее'))
 async def use(message: types.Message):
     global recomendation
+    recomendation = 0
     # end = '\n'.join(exit)
     # media = types.MediaGroup()
     # media.attach_photo(types.InputFile('Рабочий стол/media/apple - 13.jpg'), 'apple - 13')
@@ -178,14 +184,16 @@ async def use(message: types.Message):
     cur = baza.cursor()
     phons_in = cur.execute("SELECT * FROM phons_in")
     # Добавление даных из базы в выводной лист
+    print(exit)
     if len(exit) == 3:
         for row in phons_in:
             if exit != []:
-                if row[4] == exit[0] and row[5] in exit[1] and row[6] in exit[2] and row[0][:5] not in delt:
+                if exit[0] in row[4] and row[5] in exit[1] and row[6] in exit[2] and row[0][:5] not in delt:
                     recomendation += 1
-                    await message.reply(f'Вам может подойти:\n\n{row[1]}\nКлассификация:'
-                                        f'{row[4]}, {row[5]}, {row[6]}\n\n', reply_markup=klava_end)
+                    await message.answer(f'{recomendation}. Вам может подойти:\n\n{row[1]}\nКлассификация: '
+                                         f'{row[4]}, {row[5]}, {row[6]}\n\n', reply_markup=klava_end)
                     await message.answer_photo(photo=f'{row[2]}')
+                    await message.answer('-')
     # Отключение от бызы данных
     baza.close()
 
@@ -195,51 +203,75 @@ async def use(message: types.Message):
     # await message.reply(f'Вам может подойти:\n\n{", ".join(delt)}\n\n'
     #                     f'{", ".join(exit)}', reply_markup=klava_end)
 
+
 @dp.message(Command('оставить_отзыв'))
 async def add_feedback(message: types.Message):
     global fd
     fd = True
     await message.answer('Пишите ваш отзыв ;)', reply_markup=klava_start)
+
+
 @dp.message(Command('add'))
 async def add(message: types.Message):
     elem = message.text.split('; ')
-    print(elem[0][5:])
-    baza = sl.connect('phons_db.db')
-    cur = baza.cursor()
-    # Добавление данных в базу
-    cur.execute(f'INSERT INTO phons_in (name, descript, url, price, clas, market, condition) '
-                f'values(?, ?, ?, ?, ?, ?, ?)',
-                (elem[0][5:], elem[1], elem[2], elem[3], elem[4], elem[5], elem[6]))
-    # Сохранение изменений в базе
-    baza.commit()
-    await message.answer('Успешно :)', reply_markup=klava_start)
+    print(elem[0][10:])
+    if elem[0][5:10] == KEY:
+        baza = sl.connect('phons_db.db')
+        cur = baza.cursor()
+        # Добавление данных в базу
+        cur.execute(f'INSERT INTO phons_in (name, descript, url, price, clas, market, condition) '
+                    f'values(?, ?, ?, ?, ?, ?, ?)',
+                    (elem[0][10:], elem[1], elem[2], elem[3], elem[4], elem[5], elem[6]))
+        # Сохранение изменений в базе
+        baza.commit()
+        await message.answer('Успешно :)', reply_markup=klava_start)
+    else:
+        await message.answer('Неправильный ключ доступа после /add :(',
+                             reply_markup=klava_start)
+
 
 @dp.message(Command('read'))
 async def read(message: types.Message):
-    await message.answer('Смотрите :)', reply_markup=klava_start)
-    baza = sl.connect('feedback_db.db')
-    cur = baza.cursor()
-    feedbacks = cur.execute("SELECT * FROM feedbacks")
-    for row in feedbacks:
-        await message.answer(row[1], reply_markup=klava_start)
+    elem = message.text.split()
+    elem.append('')
+    if elem[1] == KEY:
+        await message.answer('Смотрите :)', reply_markup=klava_start)
+        baza = sl.connect('feedback_db.db')
+        cur = baza.cursor()
+        feedbacks = cur.execute("SELECT * FROM feedbacks")
+        for row in feedbacks:
+            await message.answer(row[1], reply_markup=klava_start)
+    else:
+        await message.answer('Неправильный ключ доступа после /read :(',
+                             reply_markup=klava_start)
+
 
 @dp.message(Command('clear'))
 async def clear(message: types.Message):
-    # Подключение бызы данных
-    baza = sl.connect('feedback_db.db')
-    cur = baza.cursor()
-    # Удаление данныз в базе
-    cur.execute('DELETE from feedbacks')
-    # Сохранение изменений в базе
-    baza.commit()
-    # Отключение от бызы данных
-    baza.close()
-    await message.answer('Отзывы удаленны из БД', reply_markup=klava_start)
+    elem = message.text.split()
+    elem.append('')
+    if elem[1] == KEY:
+        # Подключение бызы данных
+        baza = sl.connect('feedback_db.db')
+        cur = baza.cursor()
+        # Удаление данныз в базе
+        cur.execute('DELETE from feedbacks')
+        # Сохранение изменений в базе
+        baza.commit()
+        # Отключение от бызы данных
+        baza.close()
+        await message.answer('Отзывы удаленны из БД', reply_markup=klava_start)
+    else:
+        await message.answer('Неправильный ключ доступа после /clear :(',
+                             reply_markup=klava_start)
+
 
 @dp.message(F.photo)
 async def foto_url(message: types.Message):
-    photo_data = message.photo[-1]
+    photo_data = f'{message.photo[-1]}'
+    photo_data = photo_data.split("'")[1]
     await message.answer(f'{photo_data}')
+
 
 @dp.message()  # декоратор для обработчика прочих сообщений
 async def echo_message(message: types.Message):
@@ -258,7 +290,6 @@ async def echo_message(message: types.Message):
         await message.answer('Спасибо! :)', reply_markup=klava_start)
     else:
         await message.answer('Максимум что вам надо писать в чат это команду /start, а остальное - кнопки:)')
-
 
 
 if __name__ == '__main__':
